requirements.txt
annotated-types==0.7.0
audioread==3.0.1
cachetools==5.5.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.0
decorator==5.1.1
google-auth==2.37.0
google-genai==0.3.0
idna==3.10
joblib==1.4.2
lazy_loader==0.4
librosa==0.10.2.post1
llvmlite==0.43.0
msgpack==1.1.0
numba==0.60.0
numpy==2.0.2
packaging==24.2
pillow==11.0.0
platformdirs==4.3.6
pooch==1.8.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.4
pydantic_core==2.27.2
requests==2.32.3
rsa==4.9
scikit-learn==1.6.0
scipy==1.14.1
soundfile==0.12.1
soxr==0.5.0.post1
threadpoolctl==3.5.0
typing_extensions==4.12.2
urllib3==2.2.3
websockets==14.1
